test: OrderTest 테스트 추가
