+++ src/test/java/kitchenpos/order/domain/OrderLineItemTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static kitchenpos.util.TestFixture.주문테이블_1_생성;
+import static kitchenpos.order.domain.OrderTest.주문_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("OrderLineItem Domain Test")
@@ -14,18 +14,11 @@ class OrderLineItemTest {
     @Test
     void changeOrderLineItemTest() {
         // given
-        Order order = Order.of(주문테이블_1_생성(), 주문_항목_리스트_생성());
+        Order order = 주문_생성();
         order.addOrderLineItem(주문_항목_생성());
 
         // then
         assertThat(order.getOrderLineItems()).hasSize(3);
-    }
-
-    public static List<OrderLineItem> 주문_항목_리스트_생성() {
-        List<OrderLineItem> list = new ArrayList<>();
-        list.add(OrderLineItem.of(1L, 1L));
-        list.add(OrderLineItem.of(2L, 1L));
-        return list;
     }
 
     public static OrderLineItem 주문_항목_생성() {
